cleaner debug code

+++ examples/OneWireHubTest/OneWireHubTest.ino
@@ -74,7 +74,7 @@ void loop()
         static float temperature = 20.0;
         temperature += 0.1;
         if (temperature > 40.0) temperature = 10.0;
-        ds18B20.settemp(10.0);
+        ds18B20.settemp(static_cast<float>(10.0));
         ds18B21.settemp(temperature);
         Serial.println(temperature);
         // DS2450
